Batchsize of meta-paths is now configurable via url. Add some comments.
